+++ test/audits/backlinks.test.js
@@ -13,6 +13,7 @@
 /* eslint-env mocha */
 
 import { createSite } from '@adobe/spacecat-shared-data-access/src/models/site.js';
+import { createSiteTopPage } from '@adobe/spacecat-shared-data-access/src/models/site-top-page.js';
 import { createConfiguration } from '@adobe/spacecat-shared-data-access/src/models/configuration.js';
 import { createOrganization } from '@adobe/spacecat-shared-data-access/src/models/organization.js';
 
@@ -27,7 +28,8 @@ chai.use(sinonChai);
 chai.use(chaiAsPromised);
 const { expect } = chai;
 
-describe('Backlinks Tests', () => {
+describe('Backlinks Tests', function () {
+  this.timeout(10000);
   let message;
   let context;
   let mockLog;
@@ -62,6 +64,26 @@ describe('Backlinks Tests', () => {
   const configuration = createConfiguration(configurationData);
 
   // site.updateAuditTypeConfig('broken-backlinks', { disabled: false });
+
+  const siteTopPage = createSiteTopPage({
+    siteId: site.getId(),
+    url: `${site.getBaseURL()}/foo.html`,
+    traffic: 1000,
+    source: 'ahrefs',
+    geo: 'global',
+    importedAt: new Date('2024-06-18').toISOString(),
+    topKeyword: '404',
+  });
+
+  const siteTopPage2 = createSiteTopPage({
+    siteId: site.getId(),
+    url: `${site.getBaseURL()}/bar.html`,
+    traffic: 500,
+    source: 'ahrefs',
+    geo: 'global',
+    importedAt: new Date('2024-06-18').toISOString(),
+    topKeyword: '429',
+  });
 
   const site2 = createSite({
     id: 'site2',
@@ -98,6 +120,12 @@ describe('Backlinks Tests', () => {
         url_to: 'https://foo.com/returns-429',
         domain_traffic: 1000,
       },
+      {
+        title: 'backlink that times out',
+        url_from: 'https://from.com/from-4',
+        url_to: 'https://foo.com/times-out',
+        domain_traffic: 500,
+      },
     ],
   };
 
@@ -105,6 +133,7 @@ describe('Backlinks Tests', () => {
     mockDataAccess = {
       getSiteByID: sinon.stub(),
       addAudit: sinon.stub(),
+      getTopPagesForSite: sinon.stub(),
       getConfiguration: sinon.stub(),
     };
 
@@ -147,6 +176,11 @@ describe('Backlinks Tests', () => {
     nock('https://foo.com')
       .get('/returns-429')
       .reply(429);
+
+    nock('https://foo.com')
+      .get('/times-out')
+      .delay(10000)
+      .reply(200);
   });
 
   afterEach(() => {
@@ -156,6 +190,7 @@ describe('Backlinks Tests', () => {
 
   it('should successfully perform an audit to detect broken backlinks, save and send the proper audit result', async () => {
     mockDataAccess.getSiteByID = sinon.stub().withArgs('site1').resolves(site);
+    mockDataAccess.getTopPagesForSite.resolves([]);
     mockDataAccess.getConfiguration = sinon.stub().resolves(configuration);
 
     nock(site.getBaseURL())
@@ -189,8 +224,47 @@ describe('Backlinks Tests', () => {
     expect(context.log.info).to.have.been.calledWith('Successfully audited site1 for broken-backlinks type audit');
   });
 
+  it('should successfully perform an audit to detect broken backlinks based on keywords from top pages', async () => {
+    mockDataAccess.getSiteByID = sinon.stub().withArgs('site1').resolves(site);
+    mockDataAccess.getTopPagesForSite.resolves([siteTopPage, siteTopPage2]);
+
+    nock(site.getBaseURL())
+      .get(/.*/)
+      .reply(200);
+
+    nock('https://ahrefs.com')
+      .get(/.*/)
+      .reply(200, auditResult);
+
+    const expectedEnhancedBacklinks = auditResult.backlinks;
+    expectedEnhancedBacklinks[0].url_suggested = 'https://bar.foo.com/foo.html';
+    expectedEnhancedBacklinks[2].url_suggested = 'https://bar.foo.com/bar.html';
+
+    const expectedMessage = {
+      type: message.type,
+      url: site.getBaseURL(),
+      auditContext: {
+        finalUrl: 'bar.foo.com',
+      },
+      auditResult: {
+        finalUrl: 'bar.foo.com',
+        brokenBacklinks: auditResult.backlinks,
+        fullAuditRef: 'https://ahrefs.com/site-explorer/broken-backlinks?select=title%2Curl_from%2Curl_to%2Ctraffic_domain&limit=50&mode=prefix&order_by=domain_rating_source%3Adesc%2Ctraffic_domain%3Adesc&target=bar.foo.com&output=json&where=%7B%22and%22%3A%5B%7B%22field%22%3A%22is_dofollow%22%2C%22is%22%3A%5B%22eq%22%2C1%5D%7D%2C%7B%22field%22%3A%22is_content%22%2C%22is%22%3A%5B%22eq%22%2C1%5D%7D%2C%7B%22field%22%3A%22domain_rating_source%22%2C%22is%22%3A%5B%22gte%22%2C29.5%5D%7D%2C%7B%22field%22%3A%22traffic_domain%22%2C%22is%22%3A%5B%22gte%22%2C500%5D%7D%2C%7B%22field%22%3A%22links_external%22%2C%22is%22%3A%5B%22lte%22%2C300%5D%7D%5D%7D',
+      },
+    };
+
+    const response = await auditBrokenBacklinks(message, context);
+
+    expect(response.status).to.equal(204);
+    expect(mockDataAccess.addAudit).to.have.been.calledOnce;
+    expect(context.sqs.sendMessage).to.have.been.calledOnce;
+    expect(context.sqs.sendMessage).to.have.been
+      .calledWith(context.env.AUDIT_RESULTS_QUEUE_URL, expectedMessage);
+  });
+
   it('should successfully perform an audit to detect broken backlinks and set finalUrl, for baseUrl redirecting to www domain', async () => {
     mockDataAccess.getSiteByID = sinon.stub().withArgs('site2').resolves(site2);
+    mockDataAccess.getTopPagesForSite.resolves([]);
     mockDataAccess.getConfiguration = sinon.stub().resolves(configuration);
 
     nock(site2.getBaseURL())
@@ -230,8 +304,9 @@ describe('Backlinks Tests', () => {
     expect(context.log.info).to.have.been.calledWith('Successfully audited site2 for broken-backlinks type audit');
   });
 
-  it('should filter out from audit result broken backlinks the ones that return ok (even with redirection)', async () => {
+  it('should filter out from audit result broken backlinks the ones that return ok(even with redirection)', async () => {
     mockDataAccess.getSiteByID = sinon.stub().withArgs('site2').resolves(site2);
+    mockDataAccess.getTopPagesForSite.resolves([]);
     mockDataAccess.getConfiguration = sinon.stub().resolves(configuration);
 
     const fixedBacklinks = [